requirements.txt
alembic==1.5.8
amqp==5.0.6
Babel==2.9.0
bidict==0.21.2
billiard==3.6.4.0
blinker==1.4
Bootstrap-Flask==1.5.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cycler==0.10.0
dateparser==1.0.0
dnspython==2.1.0
dominate==2.6.0
elasticsearch==7.12.0
email-validator==1.1.2
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Admin==1.5.8
Flask-Babel==2.0.0
Flask-Excel==0.0.7
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Moment==0.11.0
Flask-Script==2.0.6
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
guess-language-spirit==0.5.3
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
kombu==5.0.2
lml==0.1.0
log3==0.1.7
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.4.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.16.0
python-editor==1.0.4
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
rq==1.8.0
six==1.15.0
SQLAlchemy==1.4.3
texttable==1.6.3
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
visitor==0.1.3
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
